Replaced dataObject module with GraphObject adapter
- Split off Property into its own module
- Moved old dataObject into v0 subpackage (split into two modules)
- Restored ExampleRunnerTest
