+++ PyOpenWorm/pProperty.py
@@ -1,0 +1,95 @@
+from PyOpenWorm.v0.dataObject import DataObject
+
+class Property(DataObject):
+    """ Store a value associated with a DataObject
+
+    Properties can be be accessed like methods. A method call like::
+
+        a.P()
+
+    for a property ``P`` will return values appropriate to that property for ``a``,
+    the `owner` of the property.
+
+    Parameters
+    ----------
+    owner : PyOpenWorm.dataObject.DataObject
+        The owner of this property
+    name : string
+        The name of this property. Can be accessed as an attribute like::
+
+            owner.name
+
+    """
+
+    # Indicates whether the Property is multivalued
+    multiple = False
+
+    def __init__(self, name=False, owner=False, **kwargs):
+        DataObject.__init__(self, **kwargs)
+        self.owner = owner
+        if self.owner:
+            self.owner.properties.append(self)
+            if name:
+                setattr(self.owner, name, self)
+        # XXX: Default implementation is a box for a value
+        self._value = False
+
+    def get(self,*args):
+        """ Get the things which are on the other side of this property
+
+        The return value must be iterable. For a ``get`` that just returns
+        a single value, an easy way to make an iterable is to wrap the
+        value in a tuple like ``(value,)``.
+
+        Derived classes must override.
+        """
+        # This should run a query or return a cached value
+        raise NotImplementedError()
+    def set(self,*args,**kwargs):
+        """ Set the value of this property
+
+        Derived classes must override.
+        """
+        # This should set some values and call DataObject.save()
+        raise NotImplementedError()
+
+    def one(self):
+        """ Returns a single value for the ``Property`` whether or not it is multivalued.
+        """
+
+        try:
+            r = self.get()
+            return next(iter(r))
+        except StopIteration:
+            return None
+
+    def hasValue(self):
+        """ Returns true if the Property has any values set on it.
+
+        This may be defined differently for each property
+        """
+        return True
+
+    def __call__(self,*args,**kwargs):
+        """ If arguments are passed to the ``Property``, its ``set`` method
+        is called. Otherwise, the ``get`` method is called. If the ``multiple``
+        member for the ``Property`` is set to ``True``, then a Python set containing
+        the associated values is returned. Otherwise, a single bare value is returned.
+        """
+
+        if len(args) > 0 or len(kwargs) > 0:
+            self.set(*args,**kwargs)
+            return self
+        else:
+            r = self.get(*args,**kwargs)
+            if self.multiple:
+                return set(r)
+            else:
+                try:
+                    return next(iter(r))
+                except StopIteration:
+                    return None
+
+    # Get the property (a relationship) itself
+
+
