+++ tests/MuscleTest.py
@@ -1,3 +1,4 @@
+from __future__ import print_function
 import sys
 sys.path.insert(0,".")
 import unittest
@@ -17,7 +18,7 @@ import doctest
 
 from glob import glob
 
-from GraphDBInit import * 
+from GraphDBInit import *
 
 from DataTestTemplate import _DataTest
 
@@ -31,16 +32,21 @@ class MuscleTest(_DataTest):
         n = Neuron('some neuron')
         m.innervatedBy(n)
         m.save()
-
+        print(m.rdf.serialize(format="n3"))
         v = Muscle(name='MDL08')
+        print("defined?",v.innervatedBy.defined)
+        print("fp?",v.properties)
+        print()
+        for t in v.triples():
+            print(t)
         self.assertIn(Neuron('some neuron'), list(v.innervatedBy()))
 
-    def test_muscle_neurons(self):
-        """ Should be the same as innervatedBy """
-        m = Muscle(name='MDL08')
-        neu = Neuron(name="tnnetenba")
-        m.neurons(neu)
-        m.save()
+    #def test_muscle_neurons(self):
+        #""" Should be the same as innervatedBy """
+        #m = Muscle(name='MDL08')
+        #neu = Neuron(name="tnnetenba")
+        #m.neurons(neu)
+        #m.save()
 
-        m = Muscle(name='MDL08')
-        self.assertIn(Neuron('tnnetenba'), list(m.neurons()))
+        #m = Muscle(name='MDL08')
+        #self.assertIn(Neuron('tnnetenba'), list(m.neurons()))
